refactor: remove `Creator` struct
